codegen: narrow GCC 13 sanitized mask and probe conversions
The GCC 13 Debug --sanitize build rejects 27 -Wconversion and
-Wsign-conversion diagnostics in codegen.c and in the sources machine.c
includes. Make each narrowing explicit where its bound is visible:

- u64 callee-saved and allocatable mask bits accumulated into u32
  counters: cast the completed `& 1u` addend, which is 0 or 1.
- u16 identity-plan and preserved-vector masks, whose shift promotes to
  int before `& 1u`: promote the mask itself to u32, so the shift and the
  mask stay unsigned.
- conditionals whose u8/u16 arms promote to int: narrow the completed
  conditional once. The byte probes keep their low-byte truncation, and
  the variant mask is at most 0xffff because variants_in_range bounds the
  count to 1..16.

Every changed expression yields the same value for every input, and no
warning, -Werror or sanitizer policy changes. Advances #571; #562 stays
open.

+++ src/buster/lib/compiler/codegen/codegen.c
@@ -9790,8 +9790,8 @@ BUSTER_GLOBAL_LOCAL CodegenModule codegen_generate_canonical_module_attempt(Aren
                         u32 scheduled_saved_registers = 0;
                         for (u32 physical_register = 0; physical_register < MACHINE_TARGET_REGISTER_LIMIT; physical_register += 1)
                         {
-                            placement_saved_registers += (placement.callee_saved_mask >> physical_register) & 1u;
-                            scheduled_saved_registers += (scheduled_placement.callee_saved_mask >> physical_register) & 1u;
+                            placement_saved_registers += (u32)((placement.callee_saved_mask >> physical_register) & 1u);
+                            scheduled_saved_registers += (u32)((scheduled_placement.callee_saved_mask >> physical_register) & 1u);
                         }
                         if (scheduled_placement.valid &&
                             scheduled_placement.reload_count + scheduled_placement.spill_count + 2 * scheduled_saved_registers <
@@ -9855,7 +9855,7 @@ BUSTER_GLOBAL_LOCAL CodegenModule codegen_generate_canonical_module_attempt(Aren
                             u32 machine_saved_register_count = 0;
                             for (u32 saved_register = 0; saved_register < 32u; saved_register += 1)
                             {
-                                machine_saved_register_count += (placement.callee_saved_mask >> saved_register) & 1u;
+                                machine_saved_register_count += (u32)((placement.callee_saved_mask >> saved_register) & 1u);
                             }
                             u32 machine_frame_total = placement.frame_size + 16u + 8u * machine_saved_register_count;
                             u32 machine_frame_chunks = machine_frame_total / A64_SP_ADJUST_CHUNK +
@@ -9881,7 +9881,7 @@ BUSTER_GLOBAL_LOCAL CodegenModule codegen_generate_canonical_module_attempt(Aren
                         u32 machine_push_count = 0;
                         for (u32 saved_register = 0; saved_register < 32u; saved_register += 1)
                         {
-                            machine_push_count += (placement.callee_saved_mask >> saved_register) & 1u;
+                            machine_push_count += (u32)((placement.callee_saved_mask >> saved_register) & 1u);
                         }
                         u32 machine_frame_area = placement.frame_size + 8 * machine_push_count;
                         bool machine_windows_frame = selected.function.windows_aarch64_frame;
